Fix raise hand toggled state
