+++ react/features/toolbox/functions.web.js
@@ -14,21 +14,6 @@ export { abstractMapStateToProps } from './functions.native';
 /* eslint-disable flowtype/space-before-type-colon */
 
 /**
- * Returns the button object corresponding to the given buttonName.
- *
- * @param {string} buttonName - The name of the button.
- * @param {Object} state - The current state.
- * @returns {Object} - The button object.
- */
-export function getButton(buttonName: string, state: Object) {
-    const { primaryToolbarButtons, secondaryToolbarButtons }
-        = state['features/toolbox'];
-
-    return primaryToolbarButtons.get(buttonName)
-        || secondaryToolbarButtons.get(buttonName);
-}
-
-/**
  * Takes toolbar button props and maps them to HTML attributes to set.
  *
  * @param {Object} props - Props set to the React component.
